# cedar-impex-server-application/src/main/java/org/metadatacenter/impex/resources/ImpexServerResource.java
package org.metadatacenter.impex.resources;

import com.codahale.metrics.annotation.Timed;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import com.fasterxml.jackson.databind.JsonNode;
import org.apache.commons.fileupload2.core.FileUploadException;
import org.apache.commons.fileupload2.jakarta.servlet6.JakartaServletFileUpload;
import org.metadatacenter.cadsr.form.schema.Form;
import org.metadatacenter.cadsr.ingestor.form.FormParseResult;
import org.metadatacenter.cadsr.ingestor.form.FormUtil;
import org.metadatacenter.cadsr.ingestor.util.CedarServerUtil;
import org.metadatacenter.cadsr.ingestor.util.CedarServices;
import org.metadatacenter.cadsr.ingestor.util.Constants;
import org.metadatacenter.cedar.util.dw.CedarMicroserviceResource;
import org.metadatacenter.config.CedarConfig;
import org.metadatacenter.exception.CedarException;
import org.metadatacenter.http.CedarResponseStatus;
import org.metadatacenter.impex.exception.UploadInstanceNotFoundException;
import org.metadatacenter.impex.imp.cadsr.CadsrImportStatus;
import org.metadatacenter.impex.imp.cadsr.CadsrImportStatusManager;
import org.metadatacenter.impex.imp.cadsr.CadsrImportStatusManager.ImportStatus;
import org.metadatacenter.impex.upload.FlowData;
import org.metadatacenter.impex.upload.FlowUploadUtil;
import org.metadatacenter.impex.upload.UploadManager;
import org.metadatacenter.impex.util.ImpexUtil;
import org.metadatacenter.rest.context.CedarRequestContext;
import org.metadatacenter.util.http.CedarResponse;
import org.metadatacenter.util.json.JsonMapper;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import jakarta.xml.bind.JAXBException;
import java.io.FileInputStream;
import java.io.IOException;

import static org.metadatacenter.rest.assertion.GenericAssertions.LoggedIn;

@Path("/command")
@Produces(MediaType.APPLICATION_JSON)
@Tag(name = "caDSR import")
@SecurityRequirement(name = "api_key")
public class ImpexServerResource extends CedarMicroserviceResource {

  final static Logger logger = LoggerFactory.getLogger(ImpexServerResource.class);

  public ImpexServerResource(CedarConfig cedarConfig) {
    super(cedarConfig);
  }

  /**
   * Used to upload and import caDSR forms into CEDAR. The import is executed asynchronously in an independent thread.
   * The client doesn't need to wait for the import task to be completed but it's responsible for using the
   * import-cadsr-forms-status endpoint to check the import status.
   * <p>
   * If folder Id is not provided, the imported forms will be saved into the user's home folder
   *
   * @return
   * @throws CedarException
   */
  @POST
  @Timed
  @Path("/import-cadsr-forms")
  @Consumes(MediaType.MULTIPART_FORM_DATA)
  @Operation(summary = "Upload caDSR forms and import them as templates",
      description = "Upload one or more caDSR XML forms, a chunk per request, and translate each "
          + "into a CEDAR template. The upload is chunked, so a caller sends this repeatedly; the "
          + "import starts by itself once the last chunk of an upload arrives. Translation then runs "
          + "in the background, so a 200 means the chunk was stored and any import it completed was "
          + "started, not that a template exists yet. Poll the import status to find out how it "
          + "went.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The chunk was stored; any import it completed has been started"),
      @ApiResponse(responseCode = "400", description = "The request is not a multipart upload"),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "500", description = "The chunk could not be stored")
  })
  public Response importCadsrForm(
      @Parameter(description = "Folder the templates are created in. Defaults to the caller's home folder.")
      @QueryParam("folderId") String folderId) throws CedarException {

    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);

    // Check that it's a file upload request
    if (JakartaServletFileUpload.isMultipartContent(request)) {

      try {
        String userId = c.getCedarUser().getId();
        // Extract data from the request
        FlowData data = FlowUploadUtil.getFlowData(request);
        // Every request contains a file chunk that we will save to the appropriate position of a local file
        String submissionLocalFolderPath = FlowUploadUtil
            .getUploadLocalFolderPath("impex-upload", userId, data.getUploadId());
        FlowUploadUtil.saveToLocalFile(data, userId, request.getContentLength(), submissionLocalFolderPath);

        // Update the submission upload status
        UploadManager.getInstance().updateStatus(data, userId, submissionLocalFolderPath);

        // When the upload is complete, trigger the import process
        if (UploadManager.getInstance().isUploadComplete(userId, data.getUploadId())
            && !CadsrImportStatusManager.getInstance().exists(data.getUploadId())) {

          logger.info("File(s) successfully uploaded to the Impex server: ");
          logger.info("  - Upload id: " + data.getUploadId());
          logger.info("  - Local path: " + submissionLocalFolderPath);
          logger.info("  - No. files: " + data.getTotalFilesCount());
          logger.info("  - Uploaded file names: ");
          for (String fileName : UploadManager.getInstance().getUploadFileNames(userId, data.getUploadId())) {
            logger.info("    - " + fileName);
          }

          // Import files into CEDAR asynchronously
          new Thread(() -> {
            String fileName = null;
            try {
              String cedarFolderId = folderId != null ? folderId : c.getCedarUser().getHomeFolderId();
              // Set import status to 'PENDING' for all the files that are part of the upload
              CadsrImportStatusManager.getInstance().initImportStatus(data.getUploadId(), userId, cedarFolderId);

              for (String formFilePath : UploadManager.getInstance().getUploadFilePaths(userId, data.getUploadId())) {

                try {
                  // Set status to IN_PROGRESS
                  fileName = ImpexUtil.getFileNameFromFilePath(formFilePath);
                  CadsrImportStatusManager.getInstance().setStatus(data.getUploadId(), fileName,
                      ImportStatus.IN_PROGRESS);
                  logger.info("Importing file: " + formFilePath);
                  // Translate form to CEDAR template
                  Form form = FormUtil.getForm(new FileInputStream(formFilePath));
                  FormParseResult parseResult = FormUtil.getTemplateMapFromForm(form,
                      data.getUploadId() + "_" + fileName);
                  //logger.info(JsonMapper.MAPPER.writeValueAsString(parseResult.getTemplateMap()));
                  CadsrImportStatusManager.getInstance().writeReportMessages(data.uploadId, fileName,
                      parseResult.getReportMessages());
                  // Upload template to CEDAR
                  Constants.CedarServer cedarServer = CedarServerUtil.toCedarServerFromHostName(cedarConfig.getHost());
                  String apiKey = c.getCedarUser().getFirstActiveApiKey();
                  CedarServices.createTemplate(parseResult.getTemplateMap(), cedarFolderId, cedarServer, apiKey);
                  // Set status to COMPLETE
                  CadsrImportStatusManager.getInstance().setStatus(data.getUploadId(), fileName, ImportStatus.COMPLETE);
                } catch (JAXBException e) {
                  CadsrImportStatusManager.getInstance().writeReportMessage(data.uploadId, fileName, "Parsing error: "
                    + e + "\nPlease check that the input file contains a valid XML caDSR form");
                  logger.error("Error parsing input file: " + e);
                  CadsrImportStatusManager.getInstance().setStatus(data.getUploadId(), fileName, ImportStatus.ERROR);
                } catch (IOException | RuntimeException e) {
                  CadsrImportStatusManager.getInstance().writeReportMessage(data.uploadId, fileName, "Error importing form: " + e);
                  logger.error("Error importing form: " + e);
                  CadsrImportStatusManager.getInstance().setStatus(data.getUploadId(), fileName, ImportStatus.ERROR);
                }
              }
              // Remove the upload from the status map
              UploadManager.getInstance().removeUploadStatus(userId, data.getUploadId());
            } catch (UploadInstanceNotFoundException e) {
              logger.error("Upload instance not found: " + e.getMessage());
            }
          }).start();
        }
      } catch (FileUploadException e) {
        return CedarResponse.internalServerError()
            .errorMessage("Error uploading file: " + e)
            .exception(e).build();
      } catch (IllegalAccessException e) {
        return CedarResponse.internalServerError()
            .exception(e).build();
      } catch (UploadInstanceNotFoundException e) {
        return CedarResponse.internalServerError()
            .errorMessage("Upload Id not found")
            .exception(e).build();
      } catch (IOException e) {
        return CedarResponse.internalServerError()
            .exception(e).build();
      }
      return Response.ok().build();
    } else {
      return CedarResponse.status(CedarResponseStatus.BAD_REQUEST).build();
    }
  }

  /* In production I may want to this method instead of the method below (it doesn't expose all uploads) */
//  @GET
//  @Timed
//  @Path("/import-cadsr-forms-status")
//  public Response importStatus(@QueryParam("uploadId") @NotEmpty String uploadId) throws CedarException {
//
//    CedarRequestContext c = buildRequestContext();
//    c.must(c.user()).be(LoggedIn);
//
//    try {
//      if (!CadsrImportStatusManager.getInstance().exists(uploadId)) {
//        return CedarResponse.notFound().errorMessage("The specified uploadId cannot be found").id(uploadId).build();
//      } else {
//        CadsrImportStatus status = CadsrImportStatusManager.getInstance().getStatus(uploadId);
//        JsonNode output = JsonMapper.MAPPER.valueToTree(status);
//        return Response.ok().entity(output).build();
//      }
//    } catch (Exception e) {
//      return CedarResponse.internalServerError().exception(e).build();
//    }
//  }

  @GET
  @Timed
  @Path("/import-cadsr-forms-status")
  @Operation(summary = "Get the status of a caDSR import",
      description = "Report how an import is going, file by file: pending, in progress, complete, or "
          + "in error, with the messages the translation produced. Without an upload identifier it "
          + "reports every import the caller has run. Scoped to the caller either way, and an upload "
          + "belonging to someone else is reported as not found rather than refused, so this cannot "
          + "be used to discover another user's imports.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The import status, or every status for the caller"),
      @ApiResponse(responseCode = "401", description = "Unauthorized"),
      @ApiResponse(responseCode = "404", description = "The caller has no upload with this identifier"),
      @ApiResponse(responseCode = "500", description = "Internal server error")
  })
  public Response importStatus(
      @Parameter(description = "Upload identifier. Omit it to get every import the caller has run.")
      @QueryParam("uploadId") String uploadId) throws CedarException {

    CedarRequestContext c = buildRequestContext();
    c.must(c.user()).be(LoggedIn);
    String userId = c.getCedarUser().getId();

    try {
      if (uploadId != null) {
        // Owner-scoped: not-found and not-owned are answered identically, so a user cannot probe for
        // the existence of another user's imports.
        CadsrImportStatus status = CadsrImportStatusManager.getInstance().getStatusForUser(uploadId, userId);
        if (status == null) {
          return CedarResponse.notFound().errorMessage("The specified uploadId cannot be found").id(uploadId).build();
        }
        JsonNode output = JsonMapper.MAPPER.valueToTree(status);
        return Response.ok().entity(output).build();
      } else {
        JsonNode output = JsonMapper.MAPPER.valueToTree(CadsrImportStatusManager.getInstance().getStatusesForUser(userId));
        return Response.ok().entity(output).build();
      }
    } catch (Exception e) {
      return CedarResponse.internalServerError().exception(e).build();
    }
  }
}